Telegram_Bot/requirements.txt
python-telegram-bot==20.*
httpx
